* https://gitee.com/openharmony/multimedia_audio_framework:
  repush fix unexpected invocation of setaudioscene
  回退 'Pull Request !2065 : fix unexpected invocation of setaudioscene'
  Fix the bug of AudioUtilsUnitTest and GetAudioParameter
  fix
  modify waitTime to ensure capture start success.

+++ services/audio_policy/server/include/service/session_processor.h
@@ -35,7 +35,7 @@ struct SessionEvent {
     SessionEvent() = delete;
 
     Type type;
-    uint32_t sessionID;
+    uint64_t sessionID;
     SessionInfo sessionInfo_ = {};
 };
 
@@ -43,7 +43,7 @@ class SessionProcessor {
 public:
     DISALLOW_COPY_AND_MOVE(SessionProcessor);
 
-    SessionProcessor(std::function<void(const uint32_t)> processorSessionRemoved,
+    SessionProcessor(std::function<void(const uint64_t)> processorSessionRemoved,
         std::function<void(SessionEvent)> processorSessionAdded)
         : processorSessionRemoved_(processorSessionRemoved), processorSessionAdded_(processorSessionAdded)
     {
@@ -133,7 +133,7 @@ private:
     std::mutex mutex_;
     std::condition_variable cv_;
     std::queue<SessionEvent> sessionEvents_;
-    std::function<void(const uint32_t)> processorSessionRemoved_;
+    std::function<void(const uint64_t)> processorSessionRemoved_;
     std::function<void(SessionEvent)> processorSessionAdded_;
 };
 } // namespace AudioStandard
